@@ -3,18 +3,17 @@ description: Internal data package of WebTrit Phone application.
 version: 1.0.0
 
 environment:
-  sdk: '>=2.19.0 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   clock: ^1.1.1
-  drift: ^2.4.2
-  drift_dev: ^2.4.1
+  drift: ^2.8.0
   io: ^1.0.4
   path: ^1.8.2
-  path_provider: ^2.0.0
-  sqlite3_flutter_libs: ^0.5.12
+  sqlite3_flutter_libs: ^0.5.15
 
 dev_dependencies:
-  build_runner: ^2.3.3
-  lints: ^2.0.0
-  test: ^1.22.2
+  build_runner: ^2.4.4
+  drift_dev: ^2.8.2
+  lints: ^2.1.0
+  test: ^1.24.3
