@@ -3,17 +3,17 @@ description: Library to extract/get the application info from Apple App Store an
 version: 1.0.0
 
 environment:
-  sdk: '>=2.19.2 <3.0.0'
-  flutter: ^3.7.0
+  sdk: '>=3.0.0 <4.0.0'
+  flutter: ^3.10.0
 
 dependencies:
   flutter:
     sdk: flutter
 
-  http: ^0.13.5
-  meta: ^1.8.0
-  pub_semver: ^2.1.3
-  test: ^1.22.0
+  http: ^0.13.6
+  meta: ^1.9.1
+  pub_semver: ^2.1.4
 
 dev_dependencies:
   lints: ^2.0.1
+  test: ^1.24.3
