@@ -3,18 +3,18 @@ description: Dart library for WebTrit Core API.
 version: 1.0.0
 
 environment:
-  sdk: '>=2.19.0 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   freezed_annotation: ^2.2.0
-  http: ^0.13.5
-  json_annotation: ^4.8.0
-  meta: ^1.8.0
-  pub_semver: ^2.1.3
+  http: ^0.13.6
+  json_annotation: ^4.8.1
+  meta: ^1.9.1
+  pub_semver: ^2.1.4
 
 dev_dependencies:
-  build_runner: ^2.3.3
-  freezed: ^2.3.2
-  json_serializable: ^6.6.1
-  lints: ^2.0.1
-  test: ^1.23.1
+  build_runner: ^2.4.4
+  freezed: ^2.3.4
+  json_serializable: ^6.7.0
+  lints: ^2.1.0
+  test: ^1.24.3
