@@ -3,8 +3,8 @@ description: Flutter package for processing phone numbers.
 version: 0.0.1
 
 environment:
-  sdk: '>=2.19.0 <3.0.0'
-  flutter: ^3.7.0
+  sdk: '>=3.0.0 <4.0.0'
+  flutter: ^3.10.0
 
 dependencies:
   flutter:
