@@ -3,15 +3,15 @@ description: Dart library for WebTrit Core Signalling.
 version: 1.0.0
 
 environment:
-  sdk: '>=2.19.0 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   equatable: ^2.0.5
   logging: ^1.1.1
-  logging_appenders: ^1.0.1
-  meta: ^1.8.0
-  web_socket_channel: ^2.3.0
+  logging_appenders: ^1.0.2
+  meta: ^1.9.1
+  web_socket_channel: ^2.4.0
 
 dev_dependencies:
-  lints: ^2.0.1
-  test: ^1.23.1
+  lints: ^2.1.0
+  test: ^1.24.3
