@@ -106,6 +106,8 @@ class _AppState extends State<App> {
     );
   }
 
+  final GlobalKey<NavigatorState> _mainNavigatorKey = GlobalKey<NavigatorState>(debugLabel: 'main');
+
   late final GoRouter _router = GoRouter(
     routes: [
       ShellRoute(
@@ -124,7 +126,7 @@ class _AppState extends State<App> {
             name: AppRoute.loginStep,
             path: '/login/:${LoginStep.pathParameterName}(${LoginStep.values.map((e) => e.name).join('|')})',
             builder: (context, state) {
-              final step = LoginStep.values.byName(state.params[LoginStep.pathParameterName]!);
+              final step = LoginStep.values.byName(state.pathParameters[LoginStep.pathParameterName]!);
               final widget = LoginScreen(step);
               final provider = BlocProvider(
                 create: (context) => LoginCubit(
@@ -140,7 +142,7 @@ class _AppState extends State<App> {
             path: '/web-registration',
             builder: (context, state) {
               final widget = WebRegistrationScreen(
-                initialUri: Uri.parse(state.queryParams['initialUrl'] ?? kBlankUri),
+                initialUri: Uri.parse(state.queryParameters['initialUrl'] ?? kBlankUri),
               );
               return widget;
             },
@@ -160,37 +162,183 @@ class _AppState extends State<App> {
             },
           ),
           ShellRoute(
+            navigatorKey: _mainNavigatorKey,
             builder: (context, state, child) {
               return MainShell(
-                child: child,
+                child: BlocListener<CallBloc, CallState>(
+                  listenWhen: (previous, current) => previous.activeCalls.length != current.activeCalls.length,
+                  listener: (context, state) {
+                    // TODO push/pop of call screen mechanism must be remake
+                    final router = GoRouter.of(context);
+                    final isCallLocation = router.location.startsWith(router.namedLocation(MainRoute.call));
+                    if (state.isActive) {
+                      if (!isCallLocation) {
+                        context.pushNamed(MainRoute.call);
+                        if (state.activeCall.video) {
+                          context.read<OrientationsBloc>().add(const OrientationsChanged(PreferredOrientation.call));
+                        }
+                      }
+                    } else {
+                      if (isCallLocation) {
+                        context.pop();
+                        context.read<OrientationsBloc>().add(const OrientationsChanged(PreferredOrientation.regular));
+                      }
+                    }
+                  },
+                  child: child,
+                ),
               );
             },
             routes: [
               GoRoute(
                 name: AppRoute.main,
                 path: '/main',
-                redirect: (context, state) {
-                  // propagate MainFlavor query parameter from the current router location to the new one if needed
-                  if (state.queryParams.containsKey(MainFlavor.queryParameterName)) {
-                    return null;
-                  } else {
-                    final routerLocation = Uri.parse(_router.location);
-                    final flavor =
-                        routerLocation.queryParameters[MainFlavor.queryParameterName] ?? MainFlavor.defaultValue.name;
-                    final stateLocation = Uri.parse(state.location);
-                    return stateLocation.replace(queryParameters: {
-                      ...stateLocation.queryParameters,
-                      MainFlavor.queryParameterName: flavor,
-                    }).toString();
-                  }
-                },
-                builder: (context, state) {
-                  final flavor = MainFlavor.values.byName(state.queryParams[MainFlavor.queryParameterName]!);
+                redirect: (context, state) => '/main/${MainFlavor.defaultValue.name}',
+              ),
+              StatefulShellRoute.indexedStack(
+                branches: [
+                  StatefulShellBranch(
+                    routes: [
+                      GoRoute(
+                        path: '/main/${MainFlavor.favorites.name}',
+                        name: MainRoute.favorites,
+                        builder: (context, state) {
+                          const widget = FavoritesScreen();
+                          final provider = BlocProvider(
+                            create: (context) => FavoritesBloc(
+                              favoritesRepository: context.read<FavoritesRepository>(),
+                            )..add(const FavoritesStarted()),
+                            child: widget,
+                          );
+                          return provider;
+                        },
+                        routes: [
+                          GoRoute(
+                            path: ':$contactIdPathParameterName',
+                            name: MainRoute.favoritesDetails,
+                            builder: (context, state) {
+                              const widget = ContactScreen();
+                              final provider = BlocProvider(
+                                create: (context) {
+                                  return ContactBloc(
+                                    int.parse(state.pathParameters[contactIdPathParameterName]!),
+                                    contactsRepository: context.read<ContactsRepository>(),
+                                  )..add(const ContactStarted());
+                                },
+                                child: widget,
+                              );
+                              return provider;
+                            },
+                          ),
+                        ],
+                      ),
+                    ],
+                  ),
+                  StatefulShellBranch(
+                    routes: [
+                      GoRoute(
+                        path: '/main/${MainFlavor.recents.name}',
+                        name: MainRoute.recents,
+                        builder: (context, state) {
+                          final widget = RecentsScreen(
+                            initialFilter: context.read<RecentsBloc>().state.filter,
+                          );
+                          return widget;
+                        },
+                        routes: [
+                          GoRoute(
+                            path: ':$recentIdPathParameterName',
+                            name: MainRoute.recentsDetails,
+                            builder: (context, state) {
+                              const widget = RecentScreen();
+                              var provider = BlocProvider(
+                                create: (context) {
+                                  return RecentBloc(
+                                    int.parse(state.pathParameters[recentIdPathParameterName]!),
+                                    recentsRepository: context.read<RecentsRepository>(),
+                                  )..add(const RecentStarted());
+                                },
+                                child: widget,
+                              );
+                              return provider;
+                            },
+                          ),
+                        ],
+                      ),
+                    ],
+                  ),
+                  StatefulShellBranch(
+                    routes: [
+                      GoRoute(
+                        path: '/main/${MainFlavor.contacts.name}',
+                        name: MainRoute.contacts,
+                        builder: (context, state) {
+                          final widget = ContactsScreen(
+                            sourceTypes: const [
+                              ContactSourceType.local,
+                              ContactSourceType.external,
+                            ],
+                            sourceTypeWidgetBuilder: _contactSourceTypeWidgetBuilder,
+                          );
+                          final provider = BlocProvider(
+                            create: (context) => ContactsSearchBloc(),
+                            child: widget,
+                          );
+                          return provider;
+                        },
+                        routes: [
+                          GoRoute(
+                            path: ':$contactIdPathParameterName',
+                            name: MainRoute.contactsDetails,
+                            builder: (context, state) {
+                              const widget = ContactScreen();
+                              final provider = BlocProvider(
+                                create: (context) {
+                                  return ContactBloc(
+                                    int.parse(state.pathParameters[contactIdPathParameterName]!),
+                                    contactsRepository: context.read<ContactsRepository>(),
+                                  )..add(const ContactStarted());
+                                },
+                                child: widget,
+                              );
+                              return provider;
+                            },
+                          ),
+                        ],
+                      ),
+                    ],
+                  ),
+                  StatefulShellBranch(
+                    routes: [
+                      GoRoute(
+                        path: '/main/${MainFlavor.keypad.name}',
+                        name: MainRoute.keypad,
+                        builder: (context, state) {
+                          const widget = KeypadScreen();
+                          final provider = BlocProvider(
+                            create: (context) => KeypadCubit(
+                              callBloc: context.read<CallBloc>(),
+                            ),
+                            child: widget,
+                          );
+                          return provider;
+                        },
+                      ),
+                    ],
+                  ),
+                ],
+                builder: (context, state, navigationShell) {
                   final widget = MainScreen(
-                    flavor,
-                    flavorWidgetBuilder: _flavorWidgetBuilder,
+                    navigationBarFlavor: MainFlavor.values[navigationShell.currentIndex],
+                    body: navigationShell,
+                    onNavigationBarTap: (index) {
+                      navigationShell.goBranch(
+                        index,
+                        initialLocation: index == navigationShell.currentIndex,
+                      );
+                    },
                   );
-                  final provider = BlocProvider(
+                  return BlocProvider(
                     create: (context) {
                       return MainBloc(
                         infoRepository: context.read<InfoRepository>(),
@@ -199,192 +347,135 @@ class _AppState extends State<App> {
                     },
                     child: widget,
                   );
-                  return BlocListener<CallBloc, CallState>(
-                    listenWhen: (previous, current) => previous.activeCalls.length != current.activeCalls.length,
-                    listener: (context, state) {
-                      // TODO push/pop of call screen mechanism must be remake
-                      final router = GoRouter.of(context);
-                      final isCallLocation = router.location.startsWith(router.namedLocation(MainRoute.call));
-                      if (state.isActive) {
-                        if (!isCallLocation) {
-                          context.pushNamed(MainRoute.call);
-                          if (state.activeCall.video) {
-                            context.read<OrientationsBloc>().add(const OrientationsChanged(PreferredOrientation.call));
-                          }
-                        }
-                      } else {
-                        if (isCallLocation) {
-                          context.pop();
-                          context.read<OrientationsBloc>().add(const OrientationsChanged(PreferredOrientation.regular));
-                        }
-                      }
+                },
+              ),
+              GoRoute(
+                parentNavigatorKey: _mainNavigatorKey,
+                path: '/main/call',
+                name: MainRoute.call,
+                pageBuilder: (context, state) {
+                  const widget = CallScreen();
+                  return CustomTransitionPage(
+                    key: state.pageKey,
+                    fullscreenDialog: true,
+                    child: widget,
+                    transitionsBuilder: (BuildContext context, Animation<double> animation,
+                        Animation<double> secondaryAnimation, Widget child) {
+                      const builder = ZoomPageTransitionsBuilder();
+                      return builder.buildTransitions(null, context, animation, secondaryAnimation, child);
                     },
+                  );
+                },
+              ),
+              GoRoute(
+                parentNavigatorKey: _mainNavigatorKey,
+                path: '/main/settings',
+                name: MainRoute.settings,
+                pageBuilder: (context, state) {
+                  const widget = SettingsScreen();
+                  final provider = BlocProvider(
+                    create: (context) {
+                      return SettingsBloc(
+                        appBloc: context.read<AppBloc>(),
+                        accountRepository: context.read<AccountRepository>(),
+                        appRepository: context.read<AppRepository>(),
+                        appPreferences: AppPreferences(),
+                      )..add(const SettingsRefreshed());
+                    },
+                    child: widget,
+                  );
+                  return MaterialPage(
+                    key: state.pageKey,
+                    fullscreenDialog: true,
                     child: provider,
                   );
                 },
                 routes: [
                   GoRoute(
-                    name: MainRoute.call,
-                    path: 'call',
-                    pageBuilder: (context, state) {
-                      const widget = CallScreen();
-                      return CustomTransitionPage(
-                        key: state.pageKey,
-                        fullscreenDialog: true,
-                        child: widget,
-                        transitionsBuilder: (BuildContext context, Animation<double> animation,
-                            Animation<double> secondaryAnimation, Widget child) {
-                          const builder = ZoomPageTransitionsBuilder();
-                          return builder.buildTransitions(null, context, animation, secondaryAnimation, child);
-                        },
-                      );
-                    },
-                  ),
-                  GoRoute(
-                    name: MainRoute.settings,
-                    path: 'settings',
-                    pageBuilder: (context, state) {
-                      const widget = SettingsScreen();
-                      final provider = BlocProvider(
-                        create: (context) {
-                          return SettingsBloc(
-                            appBloc: context.read<AppBloc>(),
-                            accountRepository: context.read<AccountRepository>(),
-                            appRepository: context.read<AppRepository>(),
-                            appPreferences: AppPreferences(),
-                          )..add(const SettingsRefreshed());
-                        },
-                        child: widget,
-                      );
-                      return MaterialPage(
-                        key: state.pageKey,
-                        fullscreenDialog: true,
-                        child: provider,
-                      );
-                    },
-                    routes: [
-                      GoRoute(
-                        name: MainRoute.settingsAbout,
-                        path: 'about',
-                        builder: (context, state) {
-                          if (EnvironmentConfig.APP_ABOUT_URL.isNotEmpty) {
-                            final widget = WebAboutScreen(
-                              baseAppAboutUrl: Uri.parse(EnvironmentConfig.APP_ABOUT_URL),
+                    parentNavigatorKey: _mainNavigatorKey,
+                    path: 'about',
+                    name: MainRoute.settingsAbout,
+                    builder: (context, state) {
+                      if (EnvironmentConfig.APP_ABOUT_URL.isNotEmpty) {
+                        final widget = WebAboutScreen(
+                          baseAppAboutUrl: Uri.parse(EnvironmentConfig.APP_ABOUT_URL),
+                          packageInfo: PackageInfo(),
+                          infoRepository: context.read<InfoRepository>(),
+                        );
+                        return widget;
+                      } else {
+                        const widget = AboutScreen();
+                        final provider = BlocProvider(
+                          create: (context) {
+                            return AboutBloc(
                               packageInfo: PackageInfo(),
                               infoRepository: context.read<InfoRepository>(),
-                            );
-                            return widget;
-                          } else {
-                            const widget = AboutScreen();
-                            final provider = BlocProvider(
-                              create: (context) {
-                                return AboutBloc(
-                                  packageInfo: PackageInfo(),
-                                  infoRepository: context.read<InfoRepository>(),
-                                )..add(const AboutStarted());
-                              },
-                              child: widget,
-                            );
-                            return provider;
-                          }
-                        },
-                      ),
-                      GoRoute(
-                        name: MainRoute.settingsHelp,
-                        path: 'help',
-                        builder: (context, state) {
-                          const widget = HelpScreen();
-                          return widget;
-                        },
-                      ),
-                      GoRoute(
-                        name: MainRoute.settingsLanguage,
-                        path: 'language',
-                        builder: (context, state) {
-                          const widget = LanguageScreen();
-                          return widget;
-                        },
-                      ),
-                      GoRoute(
-                        name: MainRoute.settingsNetwork,
-                        path: 'network',
-                        builder: (context, state) {
-                          const widget = NetworkScreen();
-                          return widget;
-                        },
-                      ),
-                      GoRoute(
-                        name: MainRoute.settingsTermsConditions,
-                        path: 'terms-conditions',
-                        builder: (context, state) {
-                          const widget = TermsConditionsScreen();
-                          return widget;
-                        },
-                      ),
-                      GoRoute(
-                        name: MainRoute.settingsThemeMode,
-                        path: 'theme-mode',
-                        builder: (context, state) {
-                          const widget = ThemeModeScreen();
-                          return widget;
-                        },
-                      ),
-                      GoRoute(
-                        name: MainRoute.logRecordsConsole,
-                        path: 'log-records-console',
-                        builder: (context, state) {
-                          const widget = LogRecordsConsoleScreen();
-                          final provider = BlocProvider(
-                            create: (context) => LogRecordsConsoleCubit(
-                              logRecordsRepository: context.read<LogRecordsRepository>(),
-                            )..load(),
-                            child: widget,
-                          );
-                          return provider;
-                        },
-                      ),
-                    ],
-                  ),
-                  GoRoute(
-                    name: MainRoute.contact,
-                    path: 'contact/:$contactIdPathParameterName',
-                    pageBuilder: (context, state) {
-                      const widget = ContactScreen();
-                      final provider = BlocProvider(
-                        create: (context) {
-                          return ContactBloc(
-                            int.parse(state.params[contactIdPathParameterName]!),
-                            contactsRepository: context.read<ContactsRepository>(),
-                          )..add(const ContactStarted());
-                        },
-                        child: widget,
-                      );
-                      return MaterialPage(
-                        key: state.pageKey,
-                        fullscreenDialog: true,
-                        child: provider,
-                      );
+                            )..add(const AboutStarted());
+                          },
+                          child: widget,
+                        );
+                        return provider;
+                      }
                     },
                   ),
                   GoRoute(
-                    name: MainRoute.recent,
-                    path: 'recent/:$recentIdPathParameterName',
-                    pageBuilder: (context, state) {
-                      const widget = RecentScreen();
-                      var provider = BlocProvider(
-                        create: (context) {
-                          return RecentBloc(
-                            int.parse(state.params[recentIdPathParameterName]!),
-                            recentsRepository: context.read<RecentsRepository>(),
-                          )..add(const RecentStarted());
-                        },
+                    parentNavigatorKey: _mainNavigatorKey,
+                    path: 'help',
+                    name: MainRoute.settingsHelp,
+                    builder: (context, state) {
+                      const widget = HelpScreen();
+                      return widget;
+                    },
+                  ),
+                  GoRoute(
+                    parentNavigatorKey: _mainNavigatorKey,
+                    path: 'language',
+                    name: MainRoute.settingsLanguage,
+                    builder: (context, state) {
+                      const widget = LanguageScreen();
+                      return widget;
+                    },
+                  ),
+                  GoRoute(
+                    parentNavigatorKey: _mainNavigatorKey,
+                    path: 'network',
+                    name: MainRoute.settingsNetwork,
+                    builder: (context, state) {
+                      const widget = NetworkScreen();
+                      return widget;
+                    },
+                  ),
+                  GoRoute(
+                    parentNavigatorKey: _mainNavigatorKey,
+                    path: 'terms-conditions',
+                    name: MainRoute.settingsTermsConditions,
+                    builder: (context, state) {
+                      const widget = TermsConditionsScreen();
+                      return widget;
+                    },
+                  ),
+                  GoRoute(
+                    parentNavigatorKey: _mainNavigatorKey,
+                    path: 'theme-mode',
+                    name: MainRoute.settingsThemeMode,
+                    builder: (context, state) {
+                      const widget = ThemeModeScreen();
+                      return widget;
+                    },
+                  ),
+                  GoRoute(
+                    parentNavigatorKey: _mainNavigatorKey,
+                    path: 'log-records-console',
+                    name: MainRoute.logRecordsConsole,
+                    builder: (context, state) {
+                      const widget = LogRecordsConsoleScreen();
+                      final provider = BlocProvider(
+                        create: (context) => LogRecordsConsoleCubit(
+                          logRecordsRepository: context.read<LogRecordsRepository>(),
+                        )..load(),
                         child: widget,
                       );
-                      return MaterialPage(
-                        key: state.pageKey,
-                        fullscreenDialog: true,
-                        child: provider,
-                      );
+                      return provider;
                     },
                   ),
                 ],
@@ -433,47 +524,6 @@ class _AppState extends State<App> {
     }
 
     return null;
-  }
-
-  Widget _flavorWidgetBuilder(BuildContext context, MainFlavor flavor) {
-    switch (flavor) {
-      case MainFlavor.favorites:
-        const widget = FavoritesScreen();
-        final provider = BlocProvider(
-          create: (context) => FavoritesBloc(
-            favoritesRepository: context.read<FavoritesRepository>(),
-          )..add(const FavoritesStarted()),
-          child: widget,
-        );
-        return provider;
-      case MainFlavor.recents:
-        final widget = RecentsScreen(
-          initialFilter: context.read<RecentsBloc>().state.filter,
-        );
-        return widget;
-      case MainFlavor.contacts:
-        final widget = ContactsScreen(
-          sourceTypes: const [
-            ContactSourceType.local,
-            ContactSourceType.external,
-          ],
-          sourceTypeWidgetBuilder: _contactSourceTypeWidgetBuilder,
-        );
-        final provider = BlocProvider(
-          create: (context) => ContactsSearchBloc(),
-          child: widget,
-        );
-        return provider;
-      case MainFlavor.keypad:
-        const widget = KeypadScreen();
-        final provider = BlocProvider(
-          create: (context) => KeypadCubit(
-            callBloc: context.read<CallBloc>(),
-          ),
-          child: widget,
-        );
-        return provider;
-    }
   }
 
   Widget _contactSourceTypeWidgetBuilder(BuildContext context, ContactSourceType sourceType) {
