@@ -1,15 +1,11 @@
+import 'dart:math';
+
 import 'package:flutter/material.dart';
 
 import 'package:webtrit_phone/theme/theme.dart';
 import 'package:webtrit_phone/widgets/widgets.dart';
 
 class Actionpad extends StatelessWidget {
-  final bool video;
-  final VoidCallback? onCallPressed;
-  final VoidCallback? onCallLongPress;
-  final VoidCallback? onBackspacePressed;
-  final VoidCallback? onBackspaceLongPress;
-
   const Actionpad({
     Key? key,
     this.video = false,
@@ -19,13 +15,22 @@ class Actionpad extends StatelessWidget {
     this.onBackspaceLongPress,
   }) : super(key: key);
 
+  final bool video;
+  final VoidCallback? onCallPressed;
+  final VoidCallback? onCallLongPress;
+  final VoidCallback? onBackspacePressed;
+  final VoidCallback? onBackspaceLongPress;
+
   @override
   Widget build(BuildContext context) {
     final themeData = Theme.of(context);
     final TextButtonStyles? textButtonStyles = themeData.extension<TextButtonStyles>();
 
+    final mediaQueryData = MediaQuery.of(context);
+    final minimumDimension = min(mediaQueryData.size.width / 5, mediaQueryData.size.height / 7);
+
     return TextButtonsTable(
-      minimumSize: Size.square(MediaQuery.of(context).size.width / 5),
+      minimumSize: Size.square(minimumDimension),
       children: [
         const SizedBox(),
         TextButton(
