@@ -53,7 +53,9 @@ class _LoginScreenState extends State<LoginScreen> with SingleTickerProviderStat
     return BlocListener<LoginCubit, LoginState>(
       listenWhen: (previous, current) => previous.step != current.step,
       listener: (context, state) {
-        context.goNamed(AppRoute.loginStep, params: {LoginStep.pathParameterName: state.step.name});
+        context.goNamed(AppRoute.loginStep, pathParameters: {
+          LoginStep.pathParameterName: state.step.name,
+        });
       },
       child: TabBarView(
         controller: _tabController,
