@@ -12,6 +12,7 @@ import 'package:webtrit_phone/data/data.dart';
 import 'package:webtrit_phone/environment_config.dart';
 import 'package:webtrit_phone/pre_bootstrap/pre_bootstrap.dart';
 import 'package:webtrit_phone/repositories/repositories.dart';
+import 'package:webtrit_phone/utils/utils.dart';
 
 import 'package:style/style.dart';
 
@@ -25,12 +26,15 @@ void main() {
     final logRecordsRepository = LogRecordsRepository()..attachToLogger(Logger.root);
     final appAnalyticsRepository = AppAnalyticsRepository(instance: FirebaseAnalytics.instance);
 
+    final applicationDocumentsPath = await getApplicationDocumentsPath();
+
     final theme = AppTheme().theme;
 
     return Provider<AppDatabase>(
       create: (context) {
-        final appDatabase = _AppDatabaseWithAppLifecycleStateObserver.connect(
+        final appDatabase = _AppDatabaseWithAppLifecycleStateObserver(
           createAppDatabaseConnection(
+            applicationDocumentsPath,
             'db.sqlite',
             logStatements: EnvironmentConfig.DATABASE_LOG_STATEMENTS,
           ),
@@ -63,7 +67,7 @@ void main() {
 }
 
 class _AppDatabaseWithAppLifecycleStateObserver extends AppDatabase with WidgetsBindingObserver {
-  _AppDatabaseWithAppLifecycleStateObserver.connect(connection) : super.connect(connection);
+  _AppDatabaseWithAppLifecycleStateObserver(e) : super(e);
 
   @override
   void didChangeAppLifecycleState(AppLifecycleState state) {
