@@ -1,5 +1,5 @@
 import 'package:drift/drift.dart';
 
-DatabaseConnection createAppDatabaseConnection(String name, {bool logStatements = false}) {
+DatabaseConnection createAppDatabaseConnection(String? path, String name, {bool logStatements = false}) {
   throw UnsupportedError('No suitable database connection implementation was found on this platform.');
 }
