@@ -1,7 +1,8 @@
 import 'package:drift/drift.dart';
 import 'package:drift/web.dart';
 
-DatabaseConnection createAppDatabaseConnection(String name, {bool logStatements = false}) {
+DatabaseConnection createAppDatabaseConnection(String? path, String name, {bool logStatements = false}) {
+  assert(path == null || path.isEmpty, 'path is not supported on web');
   return DatabaseConnection(
     WebDatabase(
       name,
