@@ -6,7 +6,7 @@ Future<void> main() async {
   final FlutterDriver driver = await FlutterDriver.connect();
   await integrationDriver(
     driver: driver,
-    onScreenshot: (String screenshotName, List<int> screenshotBytes) async {
+    onScreenshot: (screenshotName, screenshotBytes, [args]) async {
       // Return false if the screenshot is invalid.
       return true;
     },
